Make the 'rhdf5' package an optional dependency.

+++ tests/testthat/test_file_io.r
@@ -40,16 +40,21 @@ test_that("JSON Intermediate", {
 #----------------------------------------------------------
 # Write/Read BIOM Format 2.1 (HDFS)
 #----------------------------------------------------------
-write.biom(biom, fp, format="hdf5")
-hdf5 <- read.biom(fp)
-unlink(fp)
+if (requireNamespace("rhdf5", quietly = TRUE)) {
+  
+  write.biom(biom, fp, format="hdf5")
+  hdf5 <- read.biom(fp)
+  unlink(fp)
+  
+  test_that("HDF5 Intermediate", {
+    expect_equivalent(biom$counts,    hdf5$counts)
+    expect_equivalent(biom$phylogeny, hdf5$phylogeny)
+    expect_equivalent(biom$metadata,  hdf5$metadata)
+    expect_equivalent(biom$taxonomy,  hdf5$taxonomy)
+  })
+  
+  remove("hdf5")
+}
 
-test_that("HDF5 Intermediate", {
-  expect_equivalent(biom$counts,    hdf5$counts)
-  expect_equivalent(biom$phylogeny, hdf5$phylogeny)
-  expect_equivalent(biom$metadata,  hdf5$metadata)
-  expect_equivalent(biom$taxonomy,  hdf5$taxonomy)
-})
 
-
-remove("biom", "fp", "tsv", "json", "hdf5")
+remove("biom", "fp", "tsv", "json")
